[ADD] 특정 사원의 이력 정보 조회
1. 시작 시간이 최신 순으로 조회하도록 설정
2. 존재하지 않는 사원일 경우 400 에러 발생
3. 이력이 존재하지 않을 경우 404 에러 발생

+++ task/src/main/java/com/echoandrich/task/employee/controller/EmployeeController.java
@@ -17,7 +17,7 @@ public class EmployeeController {
     private final EmployeeService employeeService;
 
     @GetMapping(EMPLOYEE_PATH_URI)
-    public ResponseEntity<EmployeeDto> findById(@PathVariable Long employeeId) {
+    public ResponseEntity<EmployeeDto> findById(@PathVariable Integer employeeId) {
 
         EmployeeDto responseBody = employeeService.findById(employeeId);
